Moved files to standard flask placements.
